+++ config/routes.rb
@@ -2,9 +2,15 @@
 
 Rails.application.routes.draw do
   devise_for :users,
-             controllers: { omniauth_callbacks: "users/omniauth_callbacks" }
+             controllers: { omniauth_callbacks: "users/omniauth_callbacks" },
+             skip: [:sessions]
+  as :user do
+    delete "users/logout", to: "devise/sessions#destroy", as: :destroy_user_session
+  end
 
   resources :services, only: [:index, :show]
+
+  resource :profile, only: [:show]
 
   root "home#index"
 end