+++ spec/features/profile_spec.rb
@@ -1,0 +1,25 @@
+# frozen_string_literal: true
+
+require "rails_helper"
+
+RSpec.feature "Profile page" do
+  include OmniauthHelper
+
+  scenario "can be seen by authenticated user" do
+    user = create(:user)
+
+    checkin_sign_in_as(user)
+
+    click_link("My profile")
+
+    expect(page.body).to have_text(user.first_name)
+    expect(page.body).to have_text(user.last_name)
+    expect(page.body).to have_text(user.email)
+  end
+
+  scenario "link isn't visible to unauthenticated user" do
+    visit root_path
+
+    expect(page.body).to have_no_selector("nav", text: "My profile")
+  end
+end